nimbus_cli::cli

Enum CliCommand

pub(crate) enum CliCommand {

    ApplyFile {
        file: PathBuf,
        preserve_nimbus_db: bool,
        open: OpenArgs,
    },
    CaptureLogs {
        file: PathBuf,
    },
    Defaults {
        feature_id: Option<String>,
        output: Option<PathBuf>,
        manifest: ManifestArgs,
    },
    Enroll {
        experiment: ExperimentArgs,
        branch: String,
        rollouts: Vec<String>,
        preserve_targeting: bool,
        preserve_bucketing: bool,
        open: OpenArgs,
        preserve_nimbus_db: bool,
        no_validate: bool,
        manifest: ManifestArgs,
    },
    Features {
        manifest: ManifestArgs,
        experiment: ExperimentArgs,
        branch: String,
        validate: bool,
        feature_id: Option<String>,
        multi: bool,
        output: Option<PathBuf>,
    },
    Fetch {
        output: Option<PathBuf>,
        experiment: ExperimentArgs,
        recipes: Vec<String>,
    },
    FetchList {
        output: Option<PathBuf>,
        list: ExperimentListArgs,
    },
    Fml {
        args: Vec<OsString>,
    },
    Info {
        experiment: ExperimentArgs,
        output: Option<PathBuf>,
    },
    List {
        list: ExperimentListArgs,
    },
    LogState {
        open: OpenArgs,
    },
    Open {
        open: OpenArgs,
        no_clobber: bool,
    },
    StartServer,
    ResetApp,
    TailLogs,
    TestFeature {
        feature_id: String,
        files: Vec<PathBuf>,
        patch: Option<PathBuf>,
        open: OpenArgs,
        no_validate: bool,
        manifest: ManifestArgs,
    },
    Unenroll {
        open: OpenArgs,
    },
    Validate {
        experiment: ExperimentArgs,
        manifest: ManifestArgs,
    },

}

Variants

ApplyFile

Send a complete JSON file to the Nimbus SDK and apply it immediately.

Fields

file: PathBuf

The filename to be loaded into the SDK.

preserve_nimbus_db: bool

Keeps existing enrollments and experiments before enrolling.

This is unlikely what you want to do.

open: OpenArgs

CaptureLogs

Capture the logs into a file.

Fields

file: PathBuf

The file to put the logs.

Defaults

Print the defaults for the manifest.

Fields

feature_id: Option<String>

An optional feature-id

output: Option<PathBuf>

An optional file to print the manifest defaults.

manifest: ManifestArgs

Enroll

Enroll into an experiment or a rollout.

The experiment slug is a combination of the actual slug, and the server it came from.

release/stage determines the server.

preview selects the preview collection.

These can be further combined: e.g. $slug, preview/$slug, stage/$slug, stage/preview/$slug

Fields

experiment: ExperimentArgs

branch: String

The branch slug.

rollouts: Vec<String>

Optional rollout slugs, including the server and collection.

preserve_targeting: bool

Preserves the original experiment targeting

preserve_bucketing: bool

Preserves the original experiment bucketing

open: OpenArgs

preserve_nimbus_db: bool

Keeps existing enrollments and experiments before enrolling.

This is unlikely what you want to do.

no_validate: bool

Don’t validate the feature config files before enrolling

manifest: ManifestArgs

Features

Print the feature configuration involved in the branch of an experiment.

This can be optionally merged with the defaults from the feature manifest.

Fields

manifest: ManifestArgs

experiment: ExperimentArgs

branch: String

The branch of the experiment

validate: bool

If set, then merge the experimental configuration with the defaults from the manifest

feature_id: Option<String>

An optional feature-id: if it exists in this branch, print this feature on its own.

multi: bool

Print out the features involved in this branch as in a format: { $feature_id: $value }.

Automated tools should use this, since the output is predictable.

output: Option<PathBuf>

An optional file to print the output.

Fetch

Fetch one or more named experiments and rollouts and put them in a file.

Fields

output: Option<PathBuf>

The file to download the recipes to.

experiment: ExperimentArgs

recipes: Vec<String>

The recipe slugs, including server.

Use once per recipe to download. e.g. fetch –output file.json preview/my-experiment my-rollout

Cannot be used with the server option: use fetch-list instead.

FetchList

Fetch a list of experiments and put it in a file.

Fields

output: Option<PathBuf>

The file to download the recipes to.

list: ExperimentListArgs

Fml

Execute a nimbus-fml command. See

nimbus-cli fml – –help

for more.

Fields

args: Vec<OsString>

Info

Displays information about an experiment

Fields

experiment: ExperimentArgs

output: Option<PathBuf>

An optional file to print the output.

List

List the experiments from a server

Fields

list: ExperimentListArgs

LogState

Print the state of the Nimbus database to logs.

This causes a restart of the app.

Fields

open: OpenArgs

Open

Open the app without changing the state of experiment enrollments.

Fields

open: OpenArgs

no_clobber: bool

By default, the app is terminated before sending the a deeplink.

If this flag is set, then do not terminate the app if it is already running.

StartServer

Start a server

ResetApp

Reset the app back to its just installed state

TailLogs

Follow the logs for the given app.

TestFeature

Configure an application feature with one or more feature config files.

One file per branch. The branch slugs will correspond to the file names.

By default, the files are validated against the manifest; this can be overridden with --no-validate.

Fields

feature_id: String

The identifier of the feature to configure

files: Vec<PathBuf>

One or more files containing a feature config for the feature.

patch: Option<PathBuf>

An optional patch file, used to patch feature configurations

This is of the format that comes from the features --multi or defaults commands.

open: OpenArgs

no_validate: bool

Don’t validate the feature config files before enrolling

manifest: ManifestArgs

Unenroll

Unenroll from all experiments and rollouts

Fields

open: OpenArgs

Validate

Validate an experiment against a feature manifest

Fields

experiment: ExperimentArgs

manifest: ManifestArgs

Implementations

impl CliCommand

pub(crate) fn check_valid(&self) -> Result<()>

pub(crate) fn open_args(&self) -> Option<&OpenArgs>

pub(crate) fn should_kill(&self) -> bool

pub(crate) fn should_reset(&self) -> bool