nimbus_cli
Module version_utils

Functions

is_before 🔒

is_between 🔒

pad_major 🔒

pad_major will keep the string as it is. If the minor and/or patch versions are present, then they are left intact. 112.1.3 –> 112.1.3 112.1 –> 112.1 112 –> 112

pad_major_minor 🔒

pad_major_minor will zero pad the minor version if it is not present. If the patch version is present, then it is left intact. 112.1.3 –> 112.1.3 112.1 –> 112.1 112 –> 112.0

pad_major_minor_patch 🔒

The following are dumb string manipulations to pad out a version number. We might use a version library if we need much more functionality, but right now it’s isolated in a single file where it can be replaced as/when necessary.

version_split 🔒